__len__ API modification
The __len__ cannot be used any more in elf/pe: Python returns an int object,
which will cap values to 0x7FFFFFFF on 32 bit systems. A binary can have a base
address higher than this, resulting in the impossibility to handle such
programs.

The __len__ is replaced by a call to getlen.

Elfesteem needs an update to support 'max_addr' API in order to retrieve the
elf/pe max virtual address.